// src/app/layout/manage/model/client/client.ts
export class Client {
    nid_client: number=0
    scode_client: string="";
    sruc: string="";
    sname: string="";
    sbusinessname: string="";
    saddress: string="";
    sactive: string="";
    dregister: string="";
    nuserregister: number=0
    dupdate: string="";
    nuserupdate: number=0
  }